+++ source/extensions/omni.flux.properties_pane.properties.usd.widget/omni/flux/properties_pane/properties/usd/widget/setup_ui.py
@@ -43,6 +43,9 @@ class PropertyWidget:
         specific_attributes: List[str] = None,
         field_builders: list[_FieldBuilder] | None = None,
         optional_attributes: list[tuple[Callable[[Usd.Prim], bool], Dict[str, any]]] = None,
+        tree_column_widths: List[ui.Length] = None,
+        columns_resizable: bool = False,
+        right_aligned_labels: bool = True,
     ):
         """
         Property tree that show USD attributes
@@ -57,6 +60,9 @@ class PropertyWidget:
                 List of (test_func, dictionary) Tuples.
                 Test_func takes a prim and returns a bool,
                 Dictionaries should contain name, token, type, and default value.
+            tree_column_widths: the width of the columns
+            columns_resizable: if the columns are resizable
+            right_aligned_labels: if the labels are right aligned
         """
 
         self._default_attr = {
@@ -66,6 +72,9 @@ class PropertyWidget:
             "_specific_attributes": None,
             "_root_frame": None,
             "_optional_attributes": None,
+            "_tree_column_widths": None,
+            "_columns_resizable": None,
+            "_right_aligned_labels": None,
             "_property_widget": None,
             "_property_model": None,
             "_property_delegate": None,
@@ -81,6 +90,9 @@ class PropertyWidget:
         self._lookup_table = lookup_table or {}
         self._specific_attributes = specific_attributes
         self._optional_attributes = optional_attributes
+        self._tree_column_widths = tree_column_widths
+        self._columns_resizable = columns_resizable
+        self._right_aligned_labels = right_aligned_labels
 
         self.__usd_listener_instance = _get_usd_listener_instance()
 
@@ -105,7 +117,9 @@ class PropertyWidget:
 
     def __create_ui(self, field_builders: list[_FieldBuilder] | None = None):
         self._property_model = _USDPropertyModel(self._context_name)
-        self._property_delegate = _USDPropertyDelegate(field_builders=field_builders)
+        self._property_delegate = _USDPropertyDelegate(
+            field_builders=field_builders, right_aligned_labels=self._right_aligned_labels
+        )
         self._root_frame = ui.Frame(height=0)
         with self._root_frame:
             self._property_widget = _PropertyWidget(
@@ -113,6 +127,8 @@ class PropertyWidget:
                 model=self._property_model,
                 delegate=self._property_delegate,
                 refresh_callback=self.refresh,
+                tree_column_widths=self._tree_column_widths,
+                columns_resizable=self._columns_resizable,
             )
 
     def set_lookup_table(self, lookup_table: Dict[str, Dict[str, str]]):