+++ source/extensions/omni.flux.property_widget_builder.delegates/omni/flux/property_widget_builder/delegates/string_value/file_access.py
@@ -34,7 +34,7 @@ class FileAccess(AbstractField):
         text += "\nAdmin" if value & omni.client.AccessFlags.ADMIN else "\nNot admin"
         return text
 
-    def build_ui(self, item) -> list[ui.Widget]:
+    def build_ui(self, item) -> list[ui.Widget]:  # noqa PLW0221
         widgets = []
         with ui.HStack(height=ui.Pixel(20 * 3)):
             for i in range(item.element_count):