+++ source/extensions/lightspeed.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.0]
+### Added
+- Added constants for properties name column width
+
 ## [1.3.0]
 ### Added
 - Added constants for optional light parameters