+++ source/extensions/lightspeed.trex.app.style/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.9.0]
+## Added
+- Added styles for the updated properties panels ghost icons
+
 ## [1.8.5]
 ## Changed
 - Update to Kit 106.5